Rank sentences from a single document by informativity using PageRank
This uses sentence segmentation and tokenization from Chalk, a ScalaNLP
library.  Both are rule-based and have some shortcomings, but they're
a decent starting point.

Tokenized sentences are converted to vectors using Spark's TFIDF
featurization utility and used as vertices in a graph.  Sentences with
cosine similarity greater than 0.1 are connected by edges.  Running
PageRank on the graph produces sentence scores that indicate how
informative they are relative to other sentences in the document.

(Edges in the graph are weighted to make it easy to filter/prune them based on
cosine similarity, but Spark's PageRank implementation ignores the weights, so
they are effectively unweighted when computing sentence scores.)