Extract featurization into a companion object utility function
The core of the LexRank algorithm is about building a graph of related
sentences and running PageRank on the graph.  The featurization process
(which segments raw documents into sentences and vectorizes the sentences)
is more of a utility.  It would be nice to be able to summarize the same
set of documents with different parameters (e.g. edge similarity
threshold, PageRank convergence level).  Separating featurization from
the core algorithm is a first step in that direction.