Add locality-sensitive hashing for cosine similarity
Locality-sensitive hashing aims to hash similar items (i.e. those that
are close together based on some distance measure) into the same
bucket.

For LSH based on cosine similarity, a set of random vectors is
generated, each of which is a normal vector for a plane that partitions the
space.  To generate a hash for a vector of interest, dot products
with the normal vectors are computed.  The sign of each dot product
indicates which side of a particular (random) plane the vector of
interest falls, and contributes one bit to the hash.  Then, two vectors with
the same set of hash bits must fall on the same side of all of the planes, and
therefore should have low cosine distance (i.e. high cosine similarity).

Since an array of dense random vectors in a high dimensional space would
require a large amount of memory, this implementation takes advantage of a
pooling trick to reduce the number of random numbers required.  Instead
of generating a set of random vectors, a pool of random numbers is
generated.  When computing a dot product with a random vector,
components of the random vector are selected by hashing into the pool
using a combination fo the current bit number and the component index.

(See "Online Generation of Locality Sensitive Hash Signatures" by Van
Durme, et al. for a fuller explanation and justification of the pooling
trick.)