Filter out short tokens (typically punctuation and numbers)
Given the naive tokenization being used, the most common tokens are "."
and ",".  The high frequency of these punctuation tokens results in some
features that are completely uninformative and require a lot of extra
effort in the similarity computation.  (Since each feature is a matrix
row and the matrices are partitioned by row, some partitions take a long
time to compute and add very little to the overall result.)  Filtering
out these tokens reduces the skew in matrix entries across partitions,
and therefore reigns in the straggler tasks a bit.