Refactor LexRank for readability/testability and add tests
The existing code built a graph from the edges only, and later joined
the vertex information onto it.  This reworks the graph construction so
that an initial graph is constructed and subsequent operations modify
that graph.  It's easier to follow what's happening this way.  Graph
construction is also extracted to a companion object.