Add a sparse RowMatrix version of computeSimilarities to RowMatrix
An implicit comversion adds a new version of the computeSimilarities method
to RowMatrix.  The new version of the method accepts a SparseVector of
pre-computed column magnitudes, allowing the omission of columns and making
it easier to use a filtered version of a larger matrix.  This avoids out
of memory errors while computing column magnitudes with
MultivariateOnlineSummarizer, which assumes a dense matrix.