Refactor FeaturizerSuite tests to reduce duplication
An empty list of tokens or a list of tokens that are all filtered out are
equivalent and can be covered by the same test. There was also some code
duplication in setting up the list of sentences that could be removed.
These changes will make it easier to add a test for the identification
and removal of stopwords.