Identify stopwords from the corpus using TFIDF and remove them when featurizing
Select the terms with the lowest IDFs and filter them out of feature vectors as
they're created. The number of stopwords to remove can be configured with a
command-line flag, and defaults to 250.

As a side note, this also straightens out the IDF terminology to match the Spark
API. `idf` and `idfModel` were backwards; `IDFModel` is returned by `IDF.fit()`.