Remove stopword filtering from DocumentSegmenter and Configuration
Since the most common terms depend on the style and source of the
corpus, a pre-provided list of stopwords isn't very flexible. A better
way to do this would be to select stopwords from the corpus using the
inverse document frequencies of the terms. Removing the existing
stopword list and filtering code is the first step in switching over.